+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Change(tui): change config editor's theme preview elements to be in a dynamic grid.
 - Feat(tui): make the Config editor dynamically size elements.
 - Feat(tui): make the Config editor (at least rudementally) scrollable.
+- Feat(server): add extra config section for metadata scanning and parsing.
 - Feat: re-implement the Track database to allow for more search options. (Note that the old database is NOT automatically deleted)
 - Fix(tui): allow usage of key `Home`/`Pos1` in search lists.
 - Fix(tui): allow usage of key `Home`/`Pos1` and `End` in youtube search list.